requirements.txt
# ============================================================
# Dream Space: AI Interior Designer – requirements.txt
# ============================================================
#
# INSTALL ORDER:
# 1. Install PyTorch FIRST (before this file):
#    CPU:  pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#    GPU:  pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# 2. Then install this file:
#    pip install -r requirements.txt
#
# NOTE ON SCHEDULER:
#   image_generator.py now uses DPMSolverMultistepScheduler in quality mode
#   (part of diffusers >= 0.18 — already satisfied by diffusers==0.29.0 below).
#   No extra package needed.
# ============================================================

# ── Core Flask ────────────────────────────────────────────────
Flask==3.0.3
python-dotenv==1.0.1
Werkzeug==3.0.3

# ── Image processing ─────────────────────────────────────────
Pillow==10.3.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.4
scipy==1.13.0

# ── Hugging Face / Stable Diffusion ──────────────────────────
# diffusers 0.29.0 includes LCMScheduler, DDIMScheduler,
# DPMSolverMultistepScheduler, EulerAncestralDiscreteScheduler,
# StableDiffusionInpaintPipeline, StableDiffusionImg2ImgPipeline
transformers==4.41.2
diffusers==0.29.0
accelerate==0.30.1
huggingface-hub==0.23.2
safetensors==0.4.3

# ── MiDaS depth estimation ────────────────────────────────────
# MiDaS is loaded via torch.hub — no extra pip package needed.
# Requires: torch, torchvision (installed in step 1 above)
# First run will auto-download the MiDaS_small weights (~80 MB).
timm==0.9.16        # MiDaS dependency

# ── Computer vision / segmentation ───────────────────────────
ultralytics==8.2.0   # YOLOv8 segmentation model

# ── Optional: faster inference on CUDA ───────────────────────
# xformers==0.0.26   # uncomment if GPU available